+++ src/utils/render.js
@@ -1,9 +1,4 @@
-export const RenderPosition = {
-  AFTERBEGIN: 'afterbegin',
-  BEFOREEND: 'beforeend',
-  AFTEREND: 'afterend',
-  BEFOREBEGIN: 'beforebegin',
-};
+import {RenderPosition} from '../const';
 
 export const render = (container, template, position = RenderPosition.BEFOREEND) => {
   container.insertAdjacentHTML(position, template);
